+++ libaegis/input/catenate.h
@@ -1,20 +1,20 @@
 //
-//	aegis - project change supervisor
-//	Copyright (C) 2002, 2005, 2006, 2008 Peter Miller
+//      aegis - project change supervisor
+//      Copyright (C) 2002, 2005, 2006, 2008, 2012 Peter Miller
 //
-//	This program is free software; you can redistribute it and/or modify
-//	it under the terms of the GNU General Public License as published by
-//	the Free Software Foundation; either version 3 of the License, or
-//	(at your option) any later version.
+//      This program is free software; you can redistribute it and/or modify
+//      it under the terms of the GNU General Public License as published by
+//      the Free Software Foundation; either version 3 of the License, or
+//      (at your option) any later version.
 //
-//	This program is distributed in the hope that it will be useful,
-//	but WITHOUT ANY WARRANTY; without even the implied warranty of
-//	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//	GNU General Public License for more details.
+//      This program is distributed in the hope that it will be useful,
+//      but WITHOUT ANY WARRANTY; without even the implied warranty of
+//      MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//      GNU General Public License for more details.
 //
-//	You should have received a copy of the GNU General Public License
-//	along with this program. If not, see
-//	<http://www.gnu.org/licenses/>.
+//      You should have received a copy of the GNU General Public License
+//      along with this program. If not, see
+//      <http://www.gnu.org/licenses/>.
 //
 
 #ifndef LIBAEGIS_INPUT_CATENATE_H
@@ -40,16 +40,16 @@ public:
     nstring name();
 
     // See base class for documentation.
-    long length();
+    off_t length();
 
     // See base class for documentation.
     void keepalive();
 
     // See base class for documentation.
-    long read_inner(void *data, size_t nbytes);
+    ssize_t read_inner(void *data, size_t nbytes);
 
     // See base class for documentation.
-    long ftell_inner();
+    off_t ftell_inner();
 
     // See base class for documentation.
     bool is_remote() const;
@@ -59,7 +59,7 @@ private:
     size_t ndeeper;
     bool delete_on_close;
     size_t selector;
-    size_t pos;
+    off_t pos;
 
     /**
       * The default constructor.  Do not use.
@@ -78,3 +78,4 @@ private:
 };
 
 #endif // LIBAEGIS_INPUT_CATENATE_H
+// vim: set ts=8 sw=4 et :
