+++ libaegis/input/crop.h
@@ -1,20 +1,20 @@
 //
-//	aegis - project change supervisor
-//	Copyright (C) 2004-2006, 2008 Peter Miller
+//      aegis - project change supervisor
+//      Copyright (C) 2004-2006, 2008, 2012 Peter Miller
 //
-//	This program is free software; you can redistribute it and/or modify
-//	it under the terms of the GNU General Public License as published by
-//	the Free Software Foundation; either version 3 of the License, or
-//	(at your option) any later version.
+//      This program is free software; you can redistribute it and/or modify
+//      it under the terms of the GNU General Public License as published by
+//      the Free Software Foundation; either version 3 of the License, or
+//      (at your option) any later version.
 //
-//	This program is distributed in the hope that it will be useful,
-//	but WITHOUT ANY WARRANTY; without even the implied warranty of
-//	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//	GNU General Public License for more details.
+//      This program is distributed in the hope that it will be useful,
+//      but WITHOUT ANY WARRANTY; without even the implied warranty of
+//      MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//      GNU General Public License for more details.
 //
-//	You should have received a copy of the GNU General Public License
-//	along with this program. If not, see
-//	<http://www.gnu.org/licenses/>.
+//      You should have received a copy of the GNU General Public License
+//      along with this program. If not, see
+//      <http://www.gnu.org/licenses/>.
 //
 
 #ifndef LIBAEGIS_INPUT_CROP_H
@@ -43,22 +43,22 @@ public:
       * @param length
       *     How many bytes of data are in the cropped region.
       */
-    input_crop(input &deeper, long length);
+    input_crop(input &deeper, size_t length);
 
     // See base class for documentation.
     nstring name();
 
     // See base class for documentation.
-    long length();
+    off_t length();
 
     // See base class for documentation.
     void keepalive();
 
     // See base class for documentation.
-    long read_inner(void *data, size_t nbytes);
+    ssize_t read_inner(void *data, size_t nbytes);
 
     // See base class for documentation.
-    long ftell_inner();
+    off_t ftell_inner();
 
     // See base class for documentation.
     bool is_remote() const;
@@ -73,16 +73,16 @@ private:
     input deeper;
 
     /**
-      * The deeper instance variable is used to remember how many bytes
+      * The maximum instance variable is used to remember how many bytes
       * of data are in the cropped region.
       */
-    long maximum;
+    size_t maximum;
 
     /**
       * The pos instance variable is used to remember where we are up to
       * in the crop region.
       */
-    long pos;
+    size_t pos;
 
     /**
       * The name_cache instance variable is used to remember the name of
@@ -107,3 +107,4 @@ private:
 };
 
 #endif // LIBAEGIS_INPUT_CROP_H
+// vim: set ts=8 sw=4 et :
